+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.14.0.0
 
+* Added `eraUnsupportedLanguage`
+* Changed return type of `mkPlutusScript` and `mkBinaryPlutusScript` from `Maybe` to `MonadFail`
 * Deprecate `Alonzo.TxSeq` in favour of `Alonzo.BlockBody`. #5156
   * Rename `AlonzoTxSeq` to `AlonzoBlockBody`
 * Rename `alonzoEqTxRaw` to `alonzoTxEqRaw`